refactor(ocr): colocate gateway unit tests and extract lifecycle integration tests
