+++ litellm-rust/crates/ai-gateway/Cargo.toml
@@ -44,5 +44,4 @@ trace-parity = ["server", "dep:tower", "litellm-core/observability"]
 
 [dev-dependencies]
 futures-channel = "0.3"
-rstest.workspace = true
 tower = { version = "0.5.3", features = ["util"] }
