+++ litellm-rust/crates/ai-gateway/tests/ocr_lifecycle.rs
@@ -1,25 +1,19 @@
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
-use litellm_core::error::Error;
-use litellm_core::http_utils::has_header;
-use litellm_core::ocr::transformation::OcrResponseHandling;
-use rstest::rstest;
-use serde_json::{Map, Value, json};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::{TcpListener, TcpStream};
-
-use super::common_utils::{ocr_provider_config, string_headers, truncate_error_body};
-use super::prepare::prepare_ocr_call;
-use super::{OcrRequest, ocr};
-use crate::integrations::custom_guardrail::{
+use litellm_ai_gateway::integrations::custom_guardrail::{
     CustomGuardrail, GuardrailContext, GuardrailDecision, GuardrailError, GuardrailEventHook,
     GuardrailFuture, GuardrailRequest,
 };
-use crate::integrations::custom_logger::{
+use litellm_ai_gateway::integrations::custom_logger::{
     CallbackTiming, CallbackValue, CustomLogger, LogFuture, ModelCallDetails,
 };
-use crate::integrations::types::RequestMetadata;
+use litellm_ai_gateway::integrations::types::RequestMetadata;
+use litellm_ai_gateway::ocr::{OcrRequest, ocr};
+use litellm_core::error::Error;
+use serde_json::{Map, Value, json};
+use tokio::io::{AsyncReadExt, AsyncWriteExt};
+use tokio::net::{TcpListener, TcpStream};
 
 async fn read_http_headers(socket: &mut TcpStream) -> String {
     let mut request = Vec::new();
@@ -238,117 +232,6 @@ fn base_ocr_request(model: &str) -> OcrRequest<'_> {
     }
 }
 
-fn request_with_format(format: &str) -> OcrRequest<'_> {
-    let mut request = base_ocr_request("mistral/mistral-ocr-latest");
-    request.optional_params = Map::from_iter([("req_format".to_string(), json!(format))]);
-    request
-}
-
-#[rstest]
-fn native_format_rejected_for_provider_without_support_as_bad_request() {
-    let prepared = prepare_ocr_call(request_with_format("native"));
-    assert!(
-        matches!(prepared.request.config, Err(Error::InvalidRequest(message)) if message.contains("not supported for provider"))
-    );
-}
-
-#[rstest]
-fn unknown_format_rejected_for_provider_without_support_as_bad_request() {
-    let prepared = prepare_ocr_call(request_with_format("raw"));
-    assert!(
-        matches!(prepared.request.config, Err(Error::InvalidRequest(message)) if message.contains("Invalid `req_format`"))
-    );
-}
-
-#[rstest]
-#[tokio::test]
-async fn reducto_file_upload_then_parse_maps_response() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let address = listener.local_addr().expect("listener has local address");
-    let server = tokio::spawn(async move {
-        let (mut upload_socket, _) = listener.accept().await.expect("accepts upload request");
-        let upload_request = read_http_request(&mut upload_socket).await;
-        let upload_body = r#"{"file_id":"reducto://uploaded.pdf"}"#;
-        let upload_response = format!(
-            "HTTP/1.1 200 OK\r\ncontent-type: application/json\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-            upload_body.len(),
-            upload_body
-        );
-        upload_socket
-            .write_all(upload_response.as_bytes())
-            .await
-            .expect("writes upload response");
-
-        let (mut parse_socket, _) = listener.accept().await.expect("accepts parse request");
-        let parse_request = read_http_request(&mut parse_socket).await;
-        let parse_body = r#"{"job_id":"job_123","usage":{"num_pages":3,"credits":3},"result":{"chunks":[{"content":"Page 1 block A","blocks":[{"content":"Page 1 block A","bbox":{"page":1},"kind":"text"}]},{"content":"Page 2 block A","blocks":[{"content":"Page 2 block A","bbox":{"page":2},"kind":"table"}]},{"content":"Page 1 block B","blocks":[{"content":"Page 1 block B","bbox":{"page":1},"kind":"text"}]},{"content":"Page 3 block A","blocks":[{"content":"Page 3 block A","bbox":{"page":3},"kind":"figure"}]}]}}"#;
-        let parse_response = format!(
-            "HTTP/1.1 200 OK\r\ncontent-type: application/json\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-            parse_body.len(),
-            parse_body
-        );
-        parse_socket
-            .write_all(parse_response.as_bytes())
-            .await
-            .expect("writes parse response");
-        (upload_request, parse_request)
-    });
-    let api_base = format!("http://{address}");
-    let mut request = base_ocr_request("reducto/parse-v3");
-    request.api_base = Some(&api_base);
-    request.api_key = None;
-    request.extra_headers = Some(Map::from_iter([
-        ("Authorization".to_string(), json!("Bearer test-key")),
-        ("x-trace-id".to_string(), json!("trace-1")),
-    ]));
-    request.document = json!({
-        "type": "document_url",
-        "document_url": "data:application/pdf;base64,JVBERi0xLjQ="
-    });
-    request.optional_params = Map::from_iter([
-        (
-            "formatting".to_string(),
-            json!({"table_output_format": "html"}),
-        ),
-        ("retrieval".to_string(), json!({"chunk_mode": "section"})),
-        ("settings".to_string(), json!({"ocr_system": "standard"})),
-    ]);
-
-    let response = ocr(request).await.expect("Reducto OCR succeeds");
-
-    assert_eq!(response["pages"].as_array().map(Vec::len), Some(3));
-    assert_eq!(
-        response["pages"][0]["markdown"],
-        "Page 1 block A\n\nPage 1 block B"
-    );
-    assert_eq!(response["pages"][1]["markdown"], "Page 2 block A");
-    assert_eq!(response["pages"][2]["markdown"], "Page 3 block A");
-    assert_eq!(response["usage_info"]["pages_processed"], 3);
-    assert_eq!(response["usage_info"]["credits"], 3);
-    assert_eq!(response["provider_native_response"]["job_id"], "job_123");
-    let (upload_request, parse_request) = server.await.expect("server task completes");
-    assert!(
-        upload_request
-            .to_ascii_lowercase()
-            .contains("authorization: bearer test-key")
-    );
-    assert!(upload_request.contains("application/pdf"));
-    assert!(upload_request.contains("%PDF-1.4"));
-    assert!(upload_request.contains("x-trace-id: trace-1"));
-    assert!(
-        parse_request
-            .to_ascii_lowercase()
-            .contains("authorization: bearer test-key")
-    );
-    assert!(parse_request.contains(r#""input":"reducto://uploaded.pdf""#));
-    assert!(parse_request.contains(r#""table_output_format":"html""#));
-    assert!(parse_request.contains(r#""chunk_mode":"section""#));
-    assert!(parse_request.contains(r#""ocr_system":"standard""#));
-}
-
-#[rstest]
 #[tokio::test]
 async fn reducto_during_call_guardrail_blocks_before_upload() {
     let listener = TcpListener::bind("127.0.0.1:0")
@@ -373,7 +256,6 @@ async fn reducto_during_call_guardrail_blocks_before_upload() {
     assert!(accepted.is_err(), "upload socket should not be touched");
 }
 
-#[rstest]
 #[tokio::test]
 async fn reducto_upload_error_body_is_truncated() {
     let listener = TcpListener::bind("127.0.0.1:0")
@@ -408,87 +290,6 @@ async fn reducto_upload_error_body_is_truncated() {
         matches!(error, Error::Http { status: 500, body } if body.chars().count() < 300 && body.ends_with("... (truncated)"))
     );
     server.await.expect("server task completes");
-}
-
-#[test]
-fn truncate_error_body_passes_short_strings_through() {
-    let body = "Unauthorized";
-    assert_eq!(truncate_error_body(body), "Unauthorized");
-}
-
-#[test]
-fn truncate_error_body_caps_long_payloads() {
-    let body = "x".repeat(306);
-    let truncated = truncate_error_body(&body);
-
-    assert!(truncated.ends_with("... (truncated)"));
-    let prefix_chars = truncated
-        .strip_suffix("... (truncated)")
-        .expect("truncated marker present")
-        .chars()
-        .count();
-    assert_eq!(prefix_chars, 256);
-}
-
-#[test]
-fn truncate_error_body_does_not_split_multibyte_chars() {
-    let body = "é".repeat(266);
-    let truncated = truncate_error_body(&body);
-    assert!(truncated.is_char_boundary(truncated.len()));
-}
-
-#[test]
-fn ocr_dispatch_supports_migrated_providers() {
-    assert!(ocr_provider_config("mistral", "mistral-ocr-latest").is_some());
-    assert!(
-        ocr_provider_config("azure_ai", "pixtral-12b-2409")
-            .expect("azure ai config resolves")
-            .requires_data_uri_document()
-    );
-    assert_eq!(
-        ocr_provider_config("azure_ai", "doc-intelligence/prebuilt-read")
-            .expect("document intelligence config resolves")
-            .response_handling(),
-        OcrResponseHandling::AzureDocumentIntelligencePoll
-    );
-    assert!(
-        ocr_provider_config("vertex_ai", "deepseek-ocr-maas")
-            .expect("vertex deepseek config resolves")
-            .supported_ocr_params()
-            .contains(&"temperature")
-    );
-    assert!(ocr_provider_config("openai", "gpt-4o").is_none());
-}
-
-#[test]
-fn string_headers_accepts_string_values() {
-    let headers = json!({
-        "x-trace-id": "trace-1"
-    })
-    .as_object()
-    .unwrap()
-    .clone();
-
-    assert_eq!(
-        string_headers(Some(headers)).expect("string headers accepted"),
-        vec![("x-trace-id".to_string(), "trace-1".to_string())]
-    );
-}
-
-#[test]
-fn auth_header_detection_is_case_insensitive() {
-    let headers = vec![
-        ("x-trace-id".to_string(), "trace-1".to_string()),
-        ("authorization".to_string(), "Bearer sk-test".to_string()),
-    ];
-
-    assert!(has_header(&headers, "authorization"));
-
-    let headers = vec![("Authorization".to_string(), "Bearer sk-test".to_string())];
-    assert!(has_header(&headers, "authorization"));
-
-    let headers = vec![("x-trace-id".to_string(), "trace-1".to_string())];
-    assert!(!has_header(&headers, "authorization"));
 }
 
 #[tokio::test]
@@ -804,23 +605,5 @@ async fn document_intelligence_poll_uses_resolved_subscription_key() {
             .to_ascii_lowercase()
             .contains("ocp-apim-subscription-key: di-key"),
         "{poll_request}"
-    );
-}
-
-#[test]
-fn string_headers_rejects_non_string_values() {
-    let headers = json!({
-        "x-retry-count": 3
-    })
-    .as_object()
-    .unwrap()
-    .clone();
-
-    let err = string_headers(Some(headers)).expect_err("non-string header rejected");
-    assert_eq!(
-        err,
-        Error::InvalidRequest(
-            "OCR extra_headers.x-retry-count must be a string, got number".to_string()
-        )
     );
 }
